Fix msvc6 warning.

+++ sbbs3/js_file.c
@@ -2,13 +2,13 @@
 
 /* Synchronet JavaScript "File" Object */
 
-/* $Id: js_file.c,v 1.103 2008/01/18 06:06:54 deuce Exp $ */
+/* $Id: js_file.c,v 1.104 2008/02/09 23:13:50 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
  * @format.use-tabs true	(see http://www.synchro.net/ptsc_hdr.html)		*
  *																			*
- * Copyright 2007 Rob Swindell - http://www.synchro.net/copyright.html		*
+ * Copyright 2008 Rob Swindell - http://www.synchro.net/copyright.html		*
  *																			*
  * This program is free software; you can redistribute it and/or			*
  * modify it under the terms of the GNU General Public License				*
@@ -398,7 +398,7 @@ js_readbin(JSContext *cx, JSObject *obj, uintN argc, jsval *argv, jsval *rval)
 	size_t		count=1;
 	size_t		retlen;
 	void		*buffer=NULL;
-	int			i;
+	size_t		i;
     JSObject*	array;
     jsval       v;
